+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.7.6] - 2025-14-05
+
+### Added
+
+- lookup() now allows input in upper/lower case.
+- lookup() with new `preserve_case` parameter : Whether to preserve the letter case from input to suggestion.
+
 ## [6.7.5] - 2025-14-05
 
 ### Added